re #285 Fixed several input/output errors. Removed code pertaining to the old algorithm since it would not run with the corrected I/O, will instead copy the input as a placeholder. Note that the placeholder implementation assumes the number of components is the same in both input and output, and the number of threads = 1. Hole filling is still disabled until a stable new algorithm is in place.

+++ trunk/PlusLib/src/VolumeReconstruction/vtkPasteSliceIntoVolumeHelperCommon.h
@@ -161,7 +161,7 @@ static int vtkTrilinearInterpolation(F *point, T *inPtr, T *outPtr,
         }
         inPtrTmp = inPtr;
         outPtrTmp = outPtr+idx[j];
-        accPtrTmp = accPtr+ ((unsigned short)(idx[j]/outInc[0]));
+        accPtrTmp = accPtr+ ((idx[j]/outInc[0])); // removed cast to unsigned short - prevented larger increments in Z direction
         f = fdx[j];
         r = F((*accPtrTmp)/255);
         a = f + r;
